Move request into own folder
