+++ src/request/getResponseData.js
@@ -1,0 +1,8 @@
+export default response => {
+  const contentType = response.headers.get('content-type');
+  if (contentType && contentType.indexOf('application/json') !== -1) {
+    return response.json();
+  }
+
+  return response.text();
+};
